Checkpoint working tree; add .gitignore, .gitattributes, workspace docs
Large accumulated working-tree state committed as a single checkpoint so
that subsequent modernization work starts from a clean base.

Line endings
------------
The tree had been rewritten LF -> CRLF by Windows tooling, producing
~565,000 lines of phantom diff across 2,382 files and burying the real
changes. Added .gitattributes (`* text=auto eol=lf`, CRLF for .cmd/.bat/
.ps1, explicit binary markers) and staged with --renormalize.

  before normalization: 2,382 files, +661,025 / -675,059
  after  normalization: 4,691 files,  +95,942 / -106,708

Ignore rules
------------
Added a root .gitignore. Previously untracked and about to be committed:
Darkeden.rar (~700 MB, over GitHub's 100 MB file limit), Darkeden/
(~1.8 GB live install), release_site/ (~1.8 GB generated update site),
build-wsl/, build-docker/, build-docker20/, _server_logs_tmp/ (~161 MB),
plus screenshots and *.out server logs. All are build products or scratch.
Database backup *.sql files are deliberately NOT ignored - they are the
record required by the DB-change discipline.

Documentation
-------------
- Added CLAUDE.md at the workspace root: real folder layout, build
  baselines, release pipeline, server ops, DB discipline, repo hygiene.
  Prior docs described client/, server/, configs/, tooling/ - none exist.
- docs/CLAUDE.md reduced to engineering principles.
- dkrix/ and dkrixserver/ CLAUDE.md refreshed against the actual trees.
- docs/MODERNIZATION.md re-measured. Phase 5 (text pipeline) is done and
  landed out of order; Phase 1 is ~1/5 done; Phases 2/3/4 untouched
  despite the branch name; Platform.h grew to 1,996 lines against a 600
  target. Added Phase -1 (verifiable builds) as the blocking prerequisite
  and a delegation model recording what can run without supervision.

Misc
----
- Renamed char_zone_rescue_20260426_$(Get-Date -Format yyyyMMdd_HHmmss).sql
  to char_zone_rescue_20260426_224700.sql (PowerShell quoting bug had put
  the literal subshell text in the filename).

Not verified by a build. No compiler for either tree was available when
this was committed; see Phase -1 in docs/MODERNIZATION.md.

+++ dkrix/Client/Packet/Gpackets/GCRankBonusInfo.cpp
@@ -28,7 +28,7 @@ GCRankBonusInfo::~GCRankBonusInfo ()
 }
 
 //--------------------------------------------------------------------------------
-// 입력스트림(버퍼)으로부터 데이타를 읽어서 패킷을 초기화한다.
+
 //--------------------------------------------------------------------------------
 void GCRankBonusInfo::read ( SocketInputStream & iStream ) 
 	 throw ( ProtocolException , Error )
@@ -49,7 +49,7 @@ void GCRankBonusInfo::read ( SocketInputStream & iStream )
 }
 
 //--------------------------------------------------------------------------------
-// 출력스트림(버퍼)으로 패킷의 바이너리 이미지를 보낸다.
+
 //--------------------------------------------------------------------------------
 void GCRankBonusInfo::write ( SocketOutputStream & oStream ) const 
      throw ( ProtocolException , Error )
@@ -59,7 +59,7 @@ void GCRankBonusInfo::write ( SocketOutputStream & oStream ) const
 	//--------------------------------------------------
 	// write pc type
 	//--------------------------------------------------
-	oStream.write( m_RankBonusInfoList.size() );
+	oStream.write( static_cast<DWORD>(m_RankBonusInfoList.size()) );
 
 	for (std::list<DWORD>::const_iterator itr = m_RankBonusInfoList.begin(); itr!= m_RankBonusInfoList.end(); itr++) 
 	{
